feat: Made textures referenced
By making textures referenced, we prevent duplicating them when not needed and allow for reuse of textures where possible.

+++ src/texture.js
@@ -18,6 +18,11 @@ class Texture {
     if (!output) throw new Error('settings property "output" required.');
     if (!context) throw new Error('settings property "context" required.');
     this.texture = texture;
+    if (texture.refs) {
+      texture.refs++;
+    } else {
+      texture.refs = 1;
+    }
     this.size = size;
     this.dimensions = dimensions;
     this.output = output;
@@ -52,7 +57,12 @@ class Texture {
    * @desc Deletes the Texture
    */
   delete() {
+    if (this._deleted) return;
     this._deleted = true;
+    if (this.texture.refs) {
+      this.texture.refs--;
+      if (this.texture.refs) return;
+    }
     return this.context.deleteTexture(this.texture);
   }
 }
